requirements.txt
# Requirements for DERPIN Field Delineation Dashboard
numpy>=1.19.0
matplotlib>=3.3.0
rasterio>=1.2.0
opencv-python>=4.5.0
shapely>=1.7.0
ultralytics>=8.0.0
requests>=2.25.0
tqdm>=4.50.0
pillow>=8.0.0
streamlit>=1.0.0
streamlit-drawable-canvas>=0.9.0
